Decided to follow the lead of the unit tests of Fixie itself.
Test result verification now uses the Redirected console from Fixie and a set of small extension methods

As a result the FixieSpecConventionTests are much cleaner now.
Yes there are hard coded method names and its not completely refactoring safe, but the unit tests will bark, so it gets recognized.

I might want to get rid of the SymbolExtensions class in the future, because it creates to much noise.