Another round of package updates.
This time the impact was bigger than expected.
With people starting to support .Net Core transitive dependencies get more fine grained.
The update of Shouldly to v2.7 caused some of that.
The other part is caused by updating Fixie itself and the need for "version_in_path" in the paket.dependencies file.