Sample domain DeviceBase tests where not executed, because the test was abstract.
Because of that some test cases where not completely implemented and I never realized that.
(The DeviceBase is, but the sample test should not be.
Several thoughts where crossing my mind at the same time and I speculated ahead. Like having a base test for all devices
It only serves to show: Think ahead, but don't speculate in your implementation)